backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.37
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
email-validator==2.2.0
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.2
